requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.10.0
